// src/types/Offer.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface Offer { dealUuid: string, count: number, offerId: bigint, offerPropositionId: bigint, localValidFrom: string, localValidTo: string, validFromUTC: string, validToUTC: string, name: string, shortName: string, description: string, CreationDateUtc: string, imageBaseName: string, }